% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpo_maxfct.R
\name{cpoMaxFact}
\alias{cpoMaxFact}
\title{Sets a maximum number of allowed factor levels, collapses all others to "collapsed"}
\usage{
cpoMaxFact(
  max.fact.no = 2147483647L,
  id,
  export = "export.default",
  affect.type = NULL,
  affect.index = integer(0),
  affect.names = character(0),
  affect.pattern = NULL,
  affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE
)
}
\description{
Sets a maximum number of allowed factor levels, collapses all others to "collapsed"
}
