% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_config.R
\name{eval_rbv2}
\alias{eval_rbv2}
\title{Run an algorithm from rbv2}
\usage{
eval_rbv2(learner, task_id, configuration, ...)
}
\description{
@param learner \code{character}\cr
Learner name e.g. "classif.svm".
@param task  \code{character}\cr
OpenML Task ID of the corresponding task.
@param configuration \code{list}\cr
Named list of hyperparameters.
@param ... \code{any}\cr
Arguments passed on to \code{eval_config}.
}
\examples{
  learner_id = "rbv2_svm"
  task_id = 3
  configuration = list("gamma" = 0.1, cost  = 10, sample.rate = .1)
  eval_rbv2(learner_id, task_id, configuration)
}
