% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_config.R
\name{eval_config}
\alias{eval_config}
\title{Run an algorithm
@param learner \code{character}\cr
Learner name e.g. "classif.svm".
@param task  \code{character}\cr
OpenML Task ID of the corresponding task.
@param configuration \code{list}\cr
Named list of hyperparameters.
@param parallel \code{integer}\cr
Number of cores to use. Set automatically if not provided by the user.
@param logfile \code{character}\cr
Logfile to write to.
@param parallel \code{list}\cr
Metrics to compute.}
\usage{
eval_config(
  learner,
  task_id,
  configuration,
  parallel = NULL,
  logfile = NULL,
  extra_metrics = list(mlr::logloss, mlr::mmce, mlr::auc, mlr::f1, mlr::timetrain,
    mlr::timepredict)
)
}
\description{
Run an algorithm
@param learner \code{character}\cr
Learner name e.g. "classif.svm".
@param task  \code{character}\cr
OpenML Task ID of the corresponding task.
@param configuration \code{list}\cr
Named list of hyperparameters.
@param parallel \code{integer}\cr
Number of cores to use. Set automatically if not provided by the user.
@param logfile \code{character}\cr
Logfile to write to.
@param parallel \code{list}\cr
Metrics to compute.
}
\examples{
  learner_id = "classif.svm"
  task_id = 3
  configuration = list("gamma" = 0.1, cost  = 10, sample.rate = .1)
  eval_config(learner_id, task_id, configuration)
}
